Added health endpoint, fixed wrong matmul type, changed get_event_loop to get_running_loop in chat.py

+++ core/inference.py
@@ -6,7 +6,10 @@ from huggingface_hub import login
 
 login(token=settings.huggingface_token)
 
-quantization_config = BitsAndBytesConfig(load_in_4bit=True)
+quantization_config = BitsAndBytesConfig(
+    load_in_4bit=True,
+    bnb_4bit_compute_dtype=torch.float16
+)
 tokenizer = AutoTokenizer.from_pretrained(settings.model_name)
 model = AutoModelForCausalLM.from_pretrained(
     settings.model_name,
